feat(initial component): add financial boundary initial components

+++ backends/Cashflow.Domain/Entities/Category.cs
@@ -6,7 +6,7 @@ namespace Cashflow.Domain.Entities;
 
 public class Category : OwnableEntity<long, long>
 {
-    public float? MaximumBudgetInvestment { get; private set; }
+    public double? MaximumBudgetInvestment { get; private set; }
     public decimal? MaximumMoneyInvestment { get; private set; }
     public string Name { get; private set; }
